feat: migrate motivation command
